pyproject.toml
[tool.poetry]
name = "conf"
version = "0.1.0"
description = ""
authors = ["Guilherme Fragnani Bez Fontana <guilhermefragnani@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.14.0"
markdown = "^3.5.2"
django-filter = "^23.5"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
django-stubs = "^4.2.7"
mypy = "^1.8.0"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "myproject.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
